feat: [NVS-285] 실패 처리 추가
